% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createfigure1.R
\encoding{UTF-8}
\name{createfigure1}
\alias{createfigure1}
\title{createfigure1
Author:Yongfeng Shi
Version: .0.0.9000
Maintainer: yongfeng shi \href{mailto:shiyf@jlu.edu.cn}{shiyf@jlu.edu.cn}}
\usage{
createfigure1(l2, l3, main = "main", xlab = "x", ylab = "y")
}
\arguments{
\item{l2}{Enter the dataframe which is numeric class. And the information of group is listed last.}

\item{l3}{ENter the data  is the group information which is characteritic class.}

\item{main, }{xlab and ylab were characteristic type, which were use to create the title and axis names.}
}
\description{
createfigure1 function
Title:createfigure1
This function allows you to convert the numeric dataframe to cumulative frequency figure.
}
\examples{
createfigure1(x1,x2, main="Late loss distribution",xlab="Late loss (mm)",ylab = "cumulative frequency (\%)")
}
\author{
Yongfeng Shi
}
\keyword{createfigure1}
